lab1/requirements.txt
pandas
matplotlib
seaborn
requests
python-dotenv
numpy